feature(GamePage/routes): add new pages
