+++ src/routes/Game/routes/Finish/index.js
@@ -1,0 +1,9 @@
+const FinishPage = () => {
+    return (
+        <div>
+            <h1 style={{textAlign: 'center'}}>FinishPage</h1>
+        </div>
+    );
+};
+
+export default FinishPage;
